backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
google-generativeai>=0.7.2
langchain>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.7.0
httpx>=0.27.0
motor>=3.4.0
pymongo>=4.6.0
python-multipart>=0.0.9
pypdf>=4.3.1
duckduckgo-search>=6.2.6
beautifulsoup4>=4.12.3
